Restore iOS support to deletePassword… and add iOS support to _errorWithCode
SecKeychainItemDelete and SecCopyErrorMessageString are Mac-only.

Also fixes a compiler error on non-ARC Mac.

Removed unused SSKeychainErrorNoPassword.

Removed platform-defined values from SSKeychainErrorCode enum. The list is misleading on Mac OS since the actual list is much longer, and it seems unnecessary on iOS since _errorWithCode: generates the descriptions and the constants are documented in SecBase.h.